requirements.txt
scikit-learn
pandas
numpy
joblib
scapy
netifaces
python-dotenv
